Cleanup and after a thorough walk through of the ShareX docs I moved the FileUploaderService back to the same file as the FileUploader. CodeFactor will probably complain about this but it must be missconfigured since I now follow the docs
